Implemented system that checks for what weapon state player is in and executes code depending on it.
Checks weapon state and does things depending on which one player is in. Added bool that can later be changed to determine if the player is in capable of performing secondary attack. Added coroutine that executes when player presses right mouse click and is capable of performing a secondary attack. Added basic variables that can be altered to change instantiated projectiles appearance and speed via the inspector. Added basic script that destroys bullets after a set amount of time. (No collision detection yet)