Moving Platforms with Player Physics.
I was able to set up a script that allows game objects to move towards waypoints (Defined by other Game Objects).  This script can be used in the future for Enemy/NPC AI too if needed.  I also set up a script that makes the Player a parent of a moving platform which allows the platform to transfer its transformation data to the player.  However, I needed to lock the characters rotation otherwise a bug would occur.  So Z rotation is locked on the player now, would we have wanted that unlocked anyway?